WIP - Sample for adding unsupported RDM
Still trying to figure out the best way to edit/write from scratch the object for saving.

Most intuitive way would be to convert xml to a hashtable and edit that, but it looks easier said than done.

At this point, the sample is usable if you don't mind messing with Pwsh xml's utilities (xml.selectnode, etc...)